frontend/latex/output: extract summary to use hook
