+++ src/packages/frontend/frame-editors/latex-editor/output.tsx
@@ -49,136 +49,7 @@ import { use_build_logs } from "./hooks";
 import { PDFControls } from "./pdf-controls";
 import { PDFJS } from "./pdfjs";
 import { BuildLogs } from "./types";
-
-const SUMMARIZE_TEX_FILES = `
-import sys
-import json
-import re
-import os
-
-def clean_latex_text(text):
-    """Remove LaTeX commands and clean up text for readability"""
-    # Remove comments
-    text = re.sub(r'%.*$', '', text, flags=re.MULTILINE)
-
-    # Remove common LaTeX commands but preserve content
-    text = re.sub(r'\\\\(title|author|section|subsection|subsubsection|chapter)\\{([^}]*)\\}', r'**\\2**', text)
-    text = re.sub(r'\\\\(emph|textit)\\{([^}]*)\\}', r'_\\2_', text)
-    text = re.sub(r'\\\\(textbf|textsc)\\{([^}]*)\\}', r'**\\2**', text)
-
-    # Remove other LaTeX commands
-    text = re.sub(r'\\\\[a-zA-Z]+\\*?\\{[^}]*\\}', '', text)
-    text = re.sub(r'\\\\[a-zA-Z]+\\*?', '', text)
-
-    # Remove LaTeX environments but keep content
-    text = re.sub(r'\\\\begin\\{[^}]*\\}', '', text)
-    text = re.sub(r'\\\\end\\{[^}]*\\}', '', text)
-
-    # Remove excessive whitespace
-    text = re.sub(r'\\n\\s*\\n', '\\n', text)
-    text = re.sub(r'\\s+', ' ', text).strip()
-
-    return text
-
-def extract_summary(filepath, home_dir):
-    """Extract a meaningful summary from a LaTeX file"""
-    if not filepath.endswith(('.tex', '.latex')):
-        return "Non-LaTeX file"
-
-    # Handle different path formats
-    if filepath.startswith('~/'):
-        # Path starts with ~/ - replace ~ with home directory
-        expanded_path = os.path.join(home_dir, filepath[2:])
-    elif os.path.isabs(filepath):
-        # Absolute path - use as is
-        expanded_path = filepath
-    else:
-        # Relative path - join with home directory
-        expanded_path = os.path.join(home_dir, filepath)
-
-    if not os.path.exists(expanded_path):
-        return f"File not found: {expanded_path}"
-
-    try:
-        with open(expanded_path, 'r', encoding='utf-8', errors='ignore') as f:
-            content = f.read()
-    except Exception as e:
-        return f"Error reading file: {str(e)}"
-
-    # Extract first meaningful content (skip documentclass, packages, etc.)
-    lines = content.split('\\n')
-    useful_lines = []
-    in_preamble = True
-    has_document_env = '\\\\begin{document}' in content
-
-    for line in lines:
-        line = line.strip()
-        if not line or line.startswith('%'):
-            continue
-
-        # Check if we're past the preamble
-        if '\\\\begin{document}' in line:
-            in_preamble = False
-            continue
-
-        # For files without \\begin{document}, treat everything as content
-        if not has_document_env:
-            in_preamble = False
-
-        if in_preamble:
-            # Extract title, author from preamble
-            if line.startswith('\\\\title{') or line.startswith('\\\\author{'):
-                useful_lines.append(line)
-        else:
-            # Extract meaningful content
-            if any(cmd in line for cmd in ['\\\\section', '\\\\subsection', '\\\\chapter', '\\\\subsubsection']):
-                useful_lines.append(line)
-            elif line and not line.startswith('\\\\') and len(line) > 3:  # Lowered threshold
-                useful_lines.append(line)
-            elif line.startswith('\\\\') and len(line) > 10:  # Include some LaTeX commands
-                useful_lines.append(line)
-
-        # Limit to first 15 useful lines
-        if len(useful_lines) >= 15:
-            break
-
-    # If we found some useful content, use it
-    if useful_lines:
-        summary_text = '\\n'.join(useful_lines[:8])  # Use more lines
-        cleaned = clean_latex_text(summary_text)
-        if cleaned and len(cleaned.strip()) > 0:
-            # Convert to single line and truncate if too long
-            cleaned = ' '.join(cleaned.split())  # Remove all newlines and extra spaces
-            if len(cleaned) > 200:
-                cleaned = cleaned[:197] + "..."
-            return cleaned
-
-    # Fallback: show raw content (first 200 chars, cleaned)
-    # Remove comments first
-    raw_content = re.sub(r'%.*$', '', content, flags=re.MULTILINE)
-    raw_content = ' '.join(raw_content.split())  # Convert to single line
-
-    if len(raw_content) > 200:
-        raw_content = raw_content[:197] + "..."
-
-    return raw_content if raw_content else "LaTeX document"
-
-def main():
-    if len(sys.argv) < 3:
-        print(json.dumps({"error": "Usage: script.py <home_dir> <file1> <file2> ..."}))
-        return
-
-    home_dir = sys.argv[1]
-    results = {}
-
-    for filepath in sys.argv[2:]:
-        results[filepath] = extract_summary(filepath, home_dir)
-
-    print(json.dumps(results, ensure_ascii=False))
-
-if __name__ == "__main__":
-    main()
-`;
+import { useFileSummaries } from "./summarize-tex";
 
 interface OutputProps {
   id: string;
@@ -199,6 +70,7 @@ type TabType = "pdf" | "contents" | "files" | "build" | "errors";
 
 interface FileListItem {
   path: string;
+  displayPath: string;
   isMain: boolean;
   summary: string;
 }
@@ -265,15 +137,12 @@ export function Output(props: OutputProps) {
   // List of LaTeX files in the project
   const switch_to_files: List<string> = useRedux([name, "switch_to_files"]);
 
-  // File summaries state with caching (1 minute max)
-  const [fileSummaries, setFileSummaries] = useState<Record<string, string>>(
-    {},
-  );
-  const [lastSummariesFetch, setLastSummariesFetch] = useState<number>(0);
-  const [summariesLoading, setSummariesLoading] = useState<boolean>(false);
-
   // Home directory - computed once since it never changes
   const [homeDir, setHomeDir] = useState<string | null>(null);
+
+  // File summaries using the custom hook
+  const { fileSummaries, summariesLoading, refreshSummaries } =
+    useFileSummaries(switch_to_files, project_id, path, homeDir, reload);
 
   // Fetch home directory once when component mounts or project_id changes
   React.useEffect(() => {
@@ -403,102 +272,6 @@ export function Output(props: OutputProps) {
     return { errors, warnings, typesetting };
   }, [build_logs, knitr]);
 
-  // Function to generate file summaries using Python script
-  const generateFileSummaries = useCallback(
-    async (forceRefresh: boolean = false) => {
-      if (!switch_to_files || switch_to_files.size === 0) return;
-
-      const now = Date.now();
-      const oneMinute = 60 * 1000;
-
-      // Only update if it's been more than 1 minute since last fetch (unless forced)
-      if (!forceRefresh && now - lastSummariesFetch < oneMinute) return;
-
-      setSummariesLoading(true);
-
-      try {
-        // Execute Python script with file list as arguments
-        const fileList = switch_to_files.toJS();
-
-        // Write Python script to temporary file to avoid command line escaping issues
-        const scriptPath = "/tmp/tex_summarizer.py";
-        await exec({
-          command: `cat > "${scriptPath}" << 'EOFPYTHON'\n${SUMMARIZE_TEX_FILES}\nEOFPYTHON`,
-          project_id,
-          path: path_split(path).head,
-          timeout: 5,
-        });
-
-        // Use the pre-fetched home directory
-        if (!homeDir) {
-          console.warn("Home directory not available yet");
-          return;
-        }
-
-        // The switch_to_files contains canonical paths relative to the project root
-        // Pass the actual home directory to the Python script
-        const result = await exec({
-          command: "python3",
-          args: [scriptPath, homeDir, ...fileList],
-          project_id,
-          path: path_split(path).head, // Run from current file's directory
-          timeout: 30, // 30 second timeout
-        });
-
-        if (result.exit_code === 0 && result.stdout) {
-          try {
-            const summaries = JSON.parse(result.stdout);
-            setFileSummaries(summaries);
-          } catch (parseError) {
-            console.warn("Failed to parse summary results:", parseError);
-            // Fallback to basic summaries
-            const fallbackSummaries: Record<string, string> = {};
-            switch_to_files.forEach((filePath) => {
-              fallbackSummaries[filePath] = "LaTeX document";
-            });
-            setFileSummaries(fallbackSummaries);
-          }
-        } else {
-          console.warn(
-            "Summary generation failed:",
-            result.stderr ?? "Unknown error",
-          );
-          // Fallback to basic summaries
-          const fallbackSummaries: Record<string, string> = {};
-          switch_to_files.forEach((filePath) => {
-            fallbackSummaries[filePath] = "LaTeX document";
-          });
-          setFileSummaries(fallbackSummaries);
-        }
-      } catch (error) {
-        console.warn("Error generating summaries:", error);
-        // Fallback to basic summaries
-        const fallbackSummaries: Record<string, string> = {};
-        switch_to_files.forEach((filePath) => {
-          fallbackSummaries[filePath] = "LaTeX document";
-        });
-        setFileSummaries(fallbackSummaries);
-      } finally {
-        setLastSummariesFetch(now);
-        setSummariesLoading(false);
-      }
-    },
-    [switch_to_files, lastSummariesFetch, reload],
-  );
-
-  // Manual refresh function that bypasses the rate limiting
-  const refreshSummaries = useCallback(
-    () => generateFileSummaries(true),
-    [generateFileSummaries],
-  );
-
-  // Generate file summaries when files change
-  React.useEffect(() => {
-    if (switch_to_files && switch_to_files.size > 1) {
-      generateFileSummaries();
-    }
-  }, [switch_to_files, generateFileSummaries]);
-
   // No automatic tab switching - let user control tabs manually
   // Errors are indicated with red exclamation icon only
 
@@ -590,11 +363,21 @@ export function Output(props: OutputProps) {
       .sort();
     const subFileCount = subFiles.size;
 
-    const listData = subFiles.toJS().map((filePath: string) => ({
-      path: filePath,
-      isMain: false,
-      summary: fileSummaries[filePath] ?? "Loading...",
-    }));
+    // Compute the common prefix to strip (directory of main file)
+    const prefix = path_split(path).head;
+    const prefixWithSlash = prefix ? prefix + "/" : "";
+
+    const listData = subFiles.toJS().map((filePath: string) => {
+      const displayPath = filePath.startsWith(prefixWithSlash)
+        ? filePath.slice(prefixWithSlash.length)
+        : filePath;
+      return {
+        path: filePath,
+        displayPath,
+        isMain: false,
+        summary: fileSummaries[filePath] ?? "Loading...",
+      };
+    });
 
     return {
       key: "files",
@@ -686,7 +469,7 @@ export function Output(props: OutputProps) {
                           fontSize: `${uiFontSize}px`,
                         }}
                       >
-                        {item.path}
+                        {item.displayPath}
                       </span>
                     }
                     description={
